+++ test/basepack_test_app/config/initializers/netzke.rb
@@ -1,4 +1,4 @@
 Netzke::Core.setup do |config|
-  # config.ext_uri = "/extjs4"
+  config.ext_uri = "http://cdn.sencha.com/ext-4.1.1a-gpl" if ENV['EXTJS_SRC'] == 'cdn'
   config.js_direct_max_retries = 0
 end