+++ test/basepack_test_app/features/step_definitions/action_column_steps.rb
@@ -1,0 +1,3 @@
+When /^I click icon "(.*?)"$/ do |action_name|
+  find("img[data-qtip='#{action_name}']").click
+end